refactor(result): separated files
