+++ src/search/result/result.go
@@ -29,23 +29,3 @@ type Result struct {
 	ImageResult   ImageResult
 	Response      *colly.Response
 }
-
-func firstN(str string, n int) string {
-	v := []rune(str)
-	if n >= len(v) {
-		return str
-	}
-	return string(v[:n])
-}
-
-func Shorten(results []Result) []Result {
-	resultsShort := make([]Result, len(results))
-	copy(resultsShort, results)
-	for i := range resultsShort {
-		if len(resultsShort[i].Description) >= 400 {
-			descShort := firstN(resultsShort[i].Description, 397)
-			resultsShort[i].Description = descShort + "..."
-		}
-	}
-	return resultsShort
-}
